🚚 (File structure): Restructure files and folders
Flatten directory structure by moving files up
